refactor: RATING_MESSAGE 모듈화
